refactor(storage): decompose service responsibilities
