+++ .buildchain/kfd/kfd-2/release-claims.json
@@ -11,7 +11,7 @@
     "version": "4.0.0-alpha.3",
     "channel": "repository-default",
     "tag": "v4.0.0-alpha.3",
-    "sourceSha": "c8fd32f901b3af15532a11599c9b34af0b0f42bc"
+    "sourceSha": "63ec57ba35a21667cbfb59a99bb3db1fae7835c8"
   },
   "claims": [
     {
