+++ framework/core/architecture/TARGETS.cmake
@@ -77,7 +77,10 @@ set(KUNGFU_STORAGE_SERVICES_SOURCE_FILES
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/manifest_catalog_projection.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/query_render.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/service.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/service_facade.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/service_lifecycle.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/service_operation_catalog.cpp"
+  "${PROJECT_SOURCE_DIR}/src/runtime/storage/service_projection.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/source_registry_projection.cpp"
   "${PROJECT_SOURCE_DIR}/src/runtime/storage/transfer_service.cpp"
 )
