+++ framework/core/architecture/review-routes.json
@@ -1,6 +1,6 @@
 {
   "schema": "kungfu.core-review-routes/v1",
-  "authorityRoot": "sha256:3de068971b2730da9bfe92569ba481edaeaa72512bf23e02fb03799f3a880d0d",
+  "authorityRoot": "sha256:d6e9aa2820bf02c0bccf381dd51b2745c5031f96d5dc96b14e4266ce4b0ecec0",
   "fallbackAccount": "@dongkeren",
   "routes": [
     {
