+++ framework/core/src/libkungfu/src/runtime/storage/service.cpp
@@ -152,7 +152,7 @@ nlohmann::json workspace_episode_layout_json(const storage_layout_result &result
           {"notes", result.notes}};
 }
 
-nlohmann::json entries_for_manifest(const nlohmann::json &manifest, const nlohmann::json &range_filter = {}) {
+nlohmann::json entries_for_manifest(const nlohmann::json &manifest, const nlohmann::json &range_filter) {
   auto entries = array_or_empty(manifest, "entries");
   if (range_filter.is_object() && !range_filter.empty()) {
     entries = yy_storage::filter_storage_manifest_entries(entries, range_filter);
@@ -1092,453 +1092,6 @@ nlohmann::json source_registry_record_json(const yijinjing::types::AcceptedRange
           {"status", verification_status_text(record.status)}};
 }
 
-class file_storage_service : public storage_service {
-public:
-  [[nodiscard]] storage_status_result status(const storage_status_request &request) const override {
-    return status_typed_impl(request);
-  }
-
-  [[nodiscard]] storage_layout_result layout(const storage_layout_request &request) const override {
-    const auto provider = provider_cache::instance().acquire(request.runtime_dir, request.provider);
-    return workspace_episode_layout_typed(request, *provider);
-  }
-
-  [[nodiscard]] storage_fsck_result fsck(const storage_fsck_request &request) const override {
-    return fsck_typed_impl(request);
-  }
-
-  [[nodiscard]] storage_repair_plan_result repair_plan(const storage_repair_plan_request &request) const override {
-    return repair_plan_typed_impl(request);
-  }
-
-  [[nodiscard]] storage_query_result query(const storage_query_request &request) const override {
-    return query_journal_projection(request);
-  }
-
-  [[nodiscard]] storage_gc_plan_result gc_plan(const storage_gc_plan_request &request) const override {
-    return gc_plan_typed_impl(request);
-  }
-
-  [[nodiscard]] storage_rebuild_index_result
-  rebuild_index(const storage_rebuild_index_request &request) const override {
-    return rebuild_index_typed_impl(request);
-  }
-
-  [[nodiscard]] storage_compact_plan_result compact_plan(const storage_compact_plan_request &request) const override {
-    return compact_plan_typed_impl(request);
-  }
-
-  [[nodiscard]] storage_export_bundle_result
-  export_bundle(const storage_export_bundle_request &request) const override {
-    return export_bundle_typed_impl(request);
-  }
-
-  [[nodiscard]] storage_import_bundle_result
-  import_bundle(const storage_import_bundle_request &request) const override {
-    return import_bundle_typed_impl(request);
-  }
-
-  [[nodiscard]] storage_verify_sync_result verify_sync(const storage_verify_sync_request &request) const override {
-    return verify_sync_typed_impl(request);
-  }
-
-  [[nodiscard]] yijinjing::types::EpisodeOpen
-  episode_begin(const storage_episode_begin_request &request) const override {
-    return yy_storage::episode_manifest_store(request.runtime_dir).begin(request.options);
-  }
-
-  [[nodiscard]] yijinjing::types::EpisodeHeartbeat
-  episode_heartbeat(const storage_episode_heartbeat_request &request) const override {
-    return yy_storage::episode_manifest_store(request.runtime_dir).heartbeat(request.options);
-  }
-
-  [[nodiscard]] yijinjing::types::EpisodeFrameAttached
-  episode_attach_frame(const storage_episode_frame_attach_request &request) const override {
-    return yy_storage::episode_manifest_store(request.runtime_dir).attach_frame(request.options);
-  }
-
-  [[nodiscard]] yijinjing::types::EpisodeRefAttached
-  episode_attach_ref(const storage_episode_ref_attach_request &request) const override {
-    return yy_storage::episode_manifest_store(request.runtime_dir).attach_ref(request.options);
-  }
-
-  [[nodiscard]] yy_storage::episode_close_write_result
-  episode_end(const storage_episode_close_request &request) const override {
-    return yy_storage::episode_manifest_store(request.runtime_dir).end(request.options);
-  }
-
-  [[nodiscard]] yy_storage::episode_close_write_result
-  episode_abort(const storage_episode_close_request &request) const override {
-    return yy_storage::episode_manifest_store(request.runtime_dir).abort(request.options);
-  }
-
-  [[nodiscard]] yy_storage::episode_recover_result
-  episode_recover(const storage_episode_recover_request &request) const override {
-    return yy_storage::episode_manifest_store(request.runtime_dir).recover(request.options);
-  }
-
-  [[nodiscard]] storage_projection_rebuild_result
-  episode_projection_rebuild(const storage_episode_projection_rebuild_request &request) const override {
-    return episode_manifest_projection(request.runtime_dir).rebuild_typed();
-  }
-
-  [[nodiscard]] storage_episode_list_result episode_list(const storage_episode_list_request &request) const override {
-    const auto fold = yy_storage::episode_manifest_store(request.runtime_dir).fold_typed_records();
-    storage_episode_list_result result{};
-    result.runtime_dir = request.runtime_dir;
-    result.unknown_record_count = static_cast<uint64_t>(fold.unknown_record_count);
-    for (auto iter = fold.episodes.rbegin(); iter != fold.episodes.rend(); ++iter) {
-      const auto &view = iter->second;
-      const auto location_uid = view.opened ? view.open.location_uid : uint32_t{0};
-      if (request.location_uid != 0 && request.location_uid != location_uid)
-        continue;
-      result.episodes.push_back(view);
-      if (request.limit != 0 && result.episodes.size() >= request.limit)
-        break;
-    }
-    return result;
-  }
-
-  [[nodiscard]] storage_episode_inspect_result
-  episode_inspect(const storage_episode_inspect_request &request) const override {
-    const auto inspected = yy_storage::episode_manifest_store(request.runtime_dir).inspect_typed(request.episode_id);
-    storage_fsck_request fsck_request{};
-    fsck_request.runtime_dir = request.runtime_dir;
-    fsck_request.scope = storage_fsck_scope::Episode;
-    fsck_request.episode_id = request.episode_id;
-    const auto fsck_result = fsck(fsck_request);
-    return {true,
-            request.runtime_dir,
-            "yijinjing-journal",
-            inspected.episode,
-            inspected.content_root,
-            inspected.causal_graph,
-            inspected.unknown_record_count,
-            fsck_result.qualification};
-  }
-
-  [[nodiscard]] yijinjing::types::SourceRegistered
-  source_register(const storage_source_register_request &request) const override {
-    return yy_storage::source_registry_store(request.runtime_dir).register_source(request.options);
-  }
-
-  [[nodiscard]] yijinjing::types::SourceHeadUpdated
-  source_update_head(const storage_source_head_update_request &request) const override {
-    return yy_storage::source_registry_store(request.runtime_dir).update_head(request.options);
-  }
-
-  [[nodiscard]] yijinjing::types::AcceptedRangeRecorded
-  source_record_accepted_range(const storage_source_accepted_range_request &request) const override {
-    return yy_storage::source_registry_store(request.runtime_dir).record_accepted_range(request.options);
-  }
-
-  [[nodiscard]] storage_source_list_result source_list(const storage_source_list_request &request) const override {
-    const auto fold = yy_storage::source_registry_store(request.runtime_dir).fold_typed_records();
-    storage_source_list_result result{};
-    result.runtime_dir = request.runtime_dir;
-    result.unknown_record_count = static_cast<uint64_t>(fold.unknown_record_count);
-    result.sources.reserve(fold.sources.size());
-    for (const auto &[source_uid, source] : fold.sources) {
-      (void)source_uid;
-      result.sources.push_back(source);
-    }
-    return result;
-  }
-
-  [[nodiscard]] storage_source_inspect_result
-  source_inspect(const storage_source_inspect_request &request) const override {
-    const auto store = yy_storage::source_registry_store(request.runtime_dir);
-    const auto source = store.inspect_typed(request.source_id);
-    if (!source.has_value())
-      throw std::invalid_argument("source not found: " + request.source_id);
-    const auto fold = store.fold_typed_records();
-    return {true, request.runtime_dir, "yijinjing-journal", *source, static_cast<uint64_t>(fold.unknown_record_count)};
-  }
-
-  [[nodiscard]] storage_source_registry_fsck_result
-  source_registry_fsck(const storage_source_registry_fsck_request &request) const override {
-    const auto journal = yy_storage::source_registry_store(request.runtime_dir).fsck_typed(request.source_id);
-    const auto projection = source_registry_projection(request.runtime_dir).verify_typed();
-    const bool projection_degraded = projection.status == "degraded";
-    return {journal.ok && !projection_degraded, !journal.ok ? "failed" : (projection_degraded ? "degraded" : "ok"),
-            journal, projection};
-  }
-
-  [[nodiscard]] storage_projection_rebuild_result
-  source_registry_rebuild(const storage_source_registry_rebuild_request &request) const override {
-    return source_registry_projection(request.runtime_dir).rebuild_typed();
-  }
-};
-
-const file_storage_service &typed_storage_service_instance() {
-  static const file_storage_service service;
-  return service;
-}
-
-const char *episode_status_text(yy_enums::EpisodeStatus status) {
-  switch (status) {
-  case yy_enums::EpisodeStatus::Open:
-    return "open";
-  case yy_enums::EpisodeStatus::Ended:
-    return "ended";
-  case yy_enums::EpisodeStatus::Aborted:
-    return "aborted";
-  case yy_enums::EpisodeStatus::Tombstoned:
-    return "tombstoned";
-  }
-  return "unknown";
-}
-
-const char *episode_ref_kind_text(yy_enums::EpisodeRefKind kind) {
-  switch (kind) {
-  case yy_enums::EpisodeRefKind::InputFrame:
-    return "input_frame";
-  case yy_enums::EpisodeRefKind::Payload:
-    return "payload";
-  case yy_enums::EpisodeRefKind::Schema:
-    return "schema";
-  case yy_enums::EpisodeRefKind::Episode:
-    return "episode";
-  }
-  return "unknown";
-}
-
-template <typename T> nlohmann::json episode_base_record_json(const char *kind, const T &record) {
-  return {{"schema", yy_storage::EPISODE_MANIFEST_SCHEMA_V1},
-          {"record_kind", kind},
-          {"schema_version", record.schema_version},
-          {"episode_id", record.episode_id},
-          {"location_uid", record.location_uid}};
-}
-
-nlohmann::json episode_record_body_json(const yijinjing::types::EpisodeOpen &record) {
-  auto row = episode_base_record_json("episode_open", record);
-  row["status"] = episode_status_text(yy_enums::EpisodeStatus::Open);
-  row["parent_episode_id"] = record.parent_episode_id;
-  row["root_trigger_frame_uid"] = record.root_trigger_frame_uid;
-  row["begin_time"] = record.begin_time;
-  row["title"] = fixed_string(record.title);
-  row["actor"] = fixed_string(record.actor);
-  row["source"] = fixed_string(record.source);
-  return row;
-}
-
-nlohmann::json episode_record_body_json(const yijinjing::types::EpisodeHeartbeat &record) {
-  auto row = episode_base_record_json("episode_heartbeat", record);
-  row["update_time"] = record.update_time;
-  row["last_frame_uid"] = record.last_frame_uid;
-  row["frame_count"] = record.frame_count;
-  row["note"] = fixed_string(record.note);
-  return row;
-}
-
-nlohmann::json episode_record_body_json(const yijinjing::types::EpisodeFrameAttached &record) {
-  auto row = episode_base_record_json("episode_frame_attached", record);
-  row["frame_uid"] = record.frame_uid;
-  row["trigger_frame_uid"] = record.trigger_frame_uid;
-  row["stream_id"] = record.stream_id;
-  row["gen_time"] = record.gen_time;
-  row["trigger_time"] = record.trigger_time;
-  row["carrier_type"] = record.carrier_type;
-  row["source"] = record.source;
-  row["dest"] = record.dest;
-  row["data_length"] = record.data_length;
-  row["integrity_version"] = record.integrity_version;
-  row["payload_checksum"] = record.payload_checksum;
-  row["frame_checksum"] = record.frame_checksum;
-  return row;
-}
-
-nlohmann::json episode_record_body_json(const yijinjing::types::EpisodeRefAttached &record) {
-  auto row = episode_base_record_json("episode_ref_attached", record);
-  row["ref_kind"] = episode_ref_kind_text(record.ref_kind);
-  row["ref_uid"] = record.ref_uid;
-  row["update_time"] = record.update_time;
-  row["ref_id"] = fixed_string(record.ref_id);
-  row["ref_hash"] = fixed_string(record.ref_hash);
-  return row;
-}
-
-nlohmann::json episode_record_body_json(const yijinjing::types::EpisodeClosed &record) {
-  auto row = episode_base_record_json("episode_closed", record);
-  row["status"] = episode_status_text(record.status);
-  row["end_time"] = record.end_time;
-  row["last_frame_uid"] = record.last_frame_uid;
-  row["frame_count"] = record.frame_count;
-  row["reason"] = fixed_string(record.reason);
-  return row;
-}
-
-nlohmann::json episode_record_body_json(const yijinjing::types::EpisodeRootCommitted &record) {
-  auto row = episode_base_record_json("episode_root_committed", record);
-  row["commit_time"] = record.commit_time;
-  row["covered_record_count"] = record.covered_record_count;
-  row["algorithm"] = fixed_string(record.algorithm);
-  row["root_value"] = fixed_string(record.root_value);
-  return row;
-}
-
-nlohmann::json render_episode_close_write_result(const yy_storage::episode_close_write_result &result) {
-  auto rendered = episode_record_body_json(result.close);
-  if (result.content_root.has_value()) {
-    rendered["content_root"] = episode_record_body_json(*result.content_root);
-  }
-  return rendered;
-}
-
-nlohmann::json render_episode_recover_result(const yy_storage::episode_recover_result &result) {
-  nlohmann::json recovered = nlohmann::json::array();
-  for (const auto &item : result.recovered)
-    recovered.push_back(render_episode_close_write_result(item));
-  nlohmann::json skipped = nlohmann::json::array();
-  for (const auto &item : result.skipped_open)
-    skipped.push_back({{"episode_id", item.episode_id}, {"location_uid", item.location_uid}});
-  return {{"ok", true},
-          {"schema", yy_storage::EPISODE_MANIFEST_SCHEMA_V1},
-          {"runtime_dir", result.runtime_dir},
-          {"authority", "yijinjing-journal"},
-          {"recovered", std::move(recovered)},
-          {"recovered_count", result.recovered.size()},
-          {"skipped_open", std::move(skipped)},
-          {"skipped_count", result.skipped_open.size()}};
-}
-
-nlohmann::json episode_record_row_json(const yy_storage::episode_manifest_record &record) {
-  auto row = std::visit(
-      [&record](const auto &body) -> nlohmann::json {
-        using body_t = std::decay_t<decltype(body)>;
-        if constexpr (std::is_same_v<body_t, yy_storage::episode_manifest_unknown_record>) {
-          return {{"schema", yy_storage::EPISODE_MANIFEST_SCHEMA_V1},
-                  {"record_kind", "unknown"},
-                  {"carrier_type", body.carrier_type},
-                  {"frame_uid", record.manifest_frame_uid},
-                  {"gen_time", record.manifest_gen_time}};
-        } else {
-          return episode_record_body_json(body);
-        }
-      },
-      record.body);
-  row["manifest_frame_uid"] = record.manifest_frame_uid;
-  row["manifest_gen_time"] = record.manifest_gen_time;
-  return row;
-}
-
-nlohmann::json render_storage_episode_inspect_records(const storage_episode_inspect_result &result) {
-  nlohmann::json records = nlohmann::json::array();
-  for (const auto &record : result.episode.records)
-    records.push_back(episode_record_row_json(record));
-  return records;
-}
-
-nlohmann::json episode_dependency_json(const yy_storage::episode_dependency &dependency) {
-  nlohmann::json rendered = {{"kind", dependency.kind}, {"role", dependency.role}, {"status", dependency.status}};
-  if (dependency.episode_id.has_value())
-    rendered["episode_id"] = *dependency.episode_id;
-  if (dependency.frame_uid.has_value())
-    rendered["frame_uid"] = *dependency.frame_uid;
-  if (dependency.dependent_frame_uid.has_value())
-    rendered["dependent_frame_uid"] = *dependency.dependent_frame_uid;
-  if (dependency.ref_uid.has_value())
-    rendered["ref_uid"] = *dependency.ref_uid;
-  if (dependency.ref_id.has_value())
-    rendered["ref_id"] = *dependency.ref_id;
-  if (dependency.ref_hash.has_value())
-    rendered["ref_hash"] = *dependency.ref_hash;
-  return rendered;
-}
-
-nlohmann::json episode_dependencies_json(const yy_storage::episode_causal_graph &graph) {
-  nlohmann::json rendered = nlohmann::json::array();
-  for (const auto &dependency : graph.dependencies)
-    rendered.push_back(episode_dependency_json(dependency));
-  return rendered;
-}
-
-nlohmann::json episode_causal_graph_json(const yy_storage::episode_causal_graph &graph) {
-  nlohmann::json edges = nlohmann::json::array();
-  for (const auto &edge : graph.edges) {
-    edges.push_back({{"kind", "frame_trigger"},
-                     {"scope", "internal"},
-                     {"from_frame_uid", edge.from_frame_uid},
-                     {"to_frame_uid", edge.to_frame_uid}});
-  }
-  return {{"schema", graph.schema},
-          {"episode_id", graph.episode_id},
-          {"frame_count", graph.frame_count},
-          {"edge_count", graph.edges.size()},
-          {"dependency_count", graph.dependencies.size()},
-          {"degraded", graph.degraded},
-          {"edges", std::move(edges)},
-          {"dependencies", episode_dependencies_json(graph)}};
-}
-
-nlohmann::json render_storage_episode_bundle_result(const storage_episode_bundle_result &result) {
-  nlohmann::json records = nlohmann::json::array();
-  for (const auto &record : result.manifest.records)
-    records.push_back(episode_record_row_json(record));
-  nlohmann::json frames = nlohmann::json::array();
-  for (const auto index : result.manifest.frame_indices)
-    frames.push_back(episode_record_row_json(result.manifest.records.at(index)));
-  nlohmann::json refs = nlohmann::json::array();
-  for (const auto index : result.manifest.ref_indices)
-    refs.push_back(episode_record_row_json(result.manifest.records.at(index)));
-  const auto dependencies = episode_dependencies_json(result.causal_graph);
-  nlohmann::json rendered = {{"schema", "kungfu.storage.episode-bundle/v1"},
-                             {"bundle_id", result.bundle_id},
-                             {"scope", "episode"},
-                             {"episode_id", result.episode_id},
-                             {"authority", "yijinjing-journal"},
-                             {"manifest", yy_storage::episode_summary_json(result.manifest)},
-                             {"causal_graph", episode_causal_graph_json(result.causal_graph)},
-                             {"records", std::move(records)},
-                             {"frames", std::move(frames)},
-                             {"refs", std::move(refs)},
-                             {"dependencies", dependencies},
-                             {"degraded", result.causal_graph.degraded},
-                             {"record_count", result.manifest.records.size()},
-                             {"frame_count", result.manifest.frame_indices.size()},
-                             {"ref_count", result.manifest.ref_indices.size()},
-                             {"dependency_count", result.causal_graph.dependencies.size()}};
-  if (!result.self_contained) {
-    return rendered;
-  }
-  // KF-ADR-019f86da-4f90-726e-b31f-ed180aa2e7a8: bundle-owned bytes use base64 only at the JSON edge.
-  rendered["self_contained"] = true;
-  nlohmann::json journals = nlohmann::json::array();
-  for (const auto &journal : result.journals) {
-    nlohmann::json frame_rows = nlohmann::json::array();
-    for (const auto &frame : journal.frames) {
-      frame_rows.push_back({{"frame_uid", frame.frame_uid},
-                            {"gen_time", frame.gen_time},
-                            {"carrier_type", frame.carrier_type},
-                            {"frame_length", frame.frame_length},
-                            {"data_length", frame.data_length},
-                            {"bytes", base64_encode(frame.bytes)}});
-    }
-    journals.push_back({{"location",
-                         {{"role", journal.role},
-                          {"namespace", journal.namespace_},
-                          {"name", journal.name},
-                          {"mode", journal.mode},
-                          {"seed", journal.seed},
-                          {"uid", journal.location_uid}}},
-                        {"dest", journal.dest},
-                        {"frames", std::move(frame_rows)}});
-  }
-  rendered["journals"] = std::move(journals);
-  nlohmann::json ref_payloads = nlohmann::json::array();
-  for (const auto &payload : result.ref_payloads) {
-    ref_payloads.push_back({{"content_namespace", payload.content_namespace},
-                            {"ref_hash", payload.ref_hash},
-                            {"byte_len", payload.byte_len},
-                            {"bytes", base64_encode(payload.bytes)}});
-  }
-  rendered["ref_payloads"] = std::move(ref_payloads);
-  rendered["material"] = {{"missing_frame_count", result.material_missing_frame_count},
-                          {"missing_ref_payload_count", result.material_missing_ref_payload_count}};
-  return rendered;
-}
-
 nlohmann::json make_request(storage_operation operation, const storage_service_options &options) {
   return {
       {"schema", RUNTIME_STORAGE_SERVICE_SCHEMA_V1},
@@ -1557,850 +1110,5 @@ nlohmann::json make_request(storage_operation operation, const storage_service_o
 }
 
 } // namespace detail
-
-using namespace detail;
-
-const storage_service &default_storage_service() { return typed_storage_service_instance(); }
-
-nlohmann::json render_storage_episode_inspect_records(const storage_episode_inspect_result &result) {
-  return detail::render_storage_episode_inspect_records(result);
-}
-
-std::string storage_fsck_scope_name(storage_fsck_scope scope) {
-  switch (scope) {
-  case storage_fsck_scope::All:
-    return "all";
-  case storage_fsck_scope::Source:
-    return "source";
-  case storage_fsck_scope::Episode:
-    return "episode";
-  }
-  return "all";
-}
-
-storage_fsck_request parse_storage_fsck_request(const storage_service_options &options) {
-  storage_fsck_request request{};
-  request.runtime_dir = options.runtime_dir;
-  request.provider = options.provider;
-  request.provider_config_source = options.provider_config_source;
-  request.source_id = options.source_id;
-  request.episode_id = options.episode_id;
-  request.verify_frames = bool_or(options.operation_options, "verify_frames", false);
-  request.scope = options.scope == "episode"
-                      ? storage_fsck_scope::Episode
-                      : (options.source_id.empty() ? storage_fsck_scope::All : storage_fsck_scope::Source);
-  return request;
-}
-
-storage_repair_plan_request parse_storage_repair_plan_request(const storage_service_options &options) {
-  const auto fsck = parse_storage_fsck_request(options);
-  return {fsck.runtime_dir, fsck.provider,   fsck.provider_config_source, fsck.scope,
-          fsck.source_id,   fsck.episode_id, fsck.verify_frames,          options.dry_run};
-}
-
-nlohmann::json render_storage_fsck_issue(const storage_fsck_issue &issue, storage_fsck_scope scope) {
-  auto rendered = std::visit(
-      [&issue, scope](const auto &detail) {
-        using detail_t = std::decay_t<decltype(detail)>;
-        if constexpr (std::is_same_v<detail_t, storage_fsck_cross_issue>) {
-          nlohmann::json row = {{"code", issue.code}};
-          if (detail.source_id.has_value())
-            row["source_id"] = *detail.source_id;
-          if (detail.path.has_value())
-            row["path"] = *detail.path;
-          if (detail.payload_hash.has_value())
-            row["payload_hash"] = *detail.payload_hash;
-          if (detail.expected.has_value())
-            row["expected"] = *detail.expected;
-          if (detail.actual.has_value())
-            row["actual"] = *detail.actual;
-          if (detail.reason.has_value())
-            row["reason"] = *detail.reason;
-          return row;
-        } else if constexpr (std::is_same_v<detail_t, yy_storage::source_registry_fsck_issue>) {
-          nlohmann::json row = {{"code", detail.code}, {"projection", "source-registry"}};
-          if (detail.source_uid.has_value())
-            row["source_uid"] = *detail.source_uid;
-          if (detail.source_id.has_value())
-            row["source_id"] = *detail.source_id;
-          if (detail.count.has_value())
-            row["count"] = *detail.count;
-          return row;
-        } else if constexpr (std::is_same_v<detail_t, yy_storage::manifest_catalog_fsck_issue>) {
-          nlohmann::json row = {{"code", detail.code}};
-          if (detail.source_id.has_value())
-            row["source_id"] = *detail.source_id;
-          if (detail.manifest_id.has_value())
-            row["manifest_id"] = *detail.manifest_id;
-          if (detail.error.has_value())
-            row["error"] = *detail.error;
-          if (detail.subject.has_value())
-            row["subject"] = *detail.subject;
-          if (detail.payload_hash.has_value())
-            row["payload_hash"] = *detail.payload_hash;
-          if (detail.state.has_value())
-            row["state"] = *detail.state;
-          if (detail.kind.has_value())
-            row["kind"] = *detail.kind;
-          if (detail.entry_source_id.has_value())
-            row["entry_source_id"] = *detail.entry_source_id;
-          if (detail.manifest_uid.has_value())
-            row["manifest_uid"] = *detail.manifest_uid;
-          if (detail.entry_index.has_value())
-            row["entry_index"] = *detail.entry_index;
-          if (detail.expected.has_value())
-            row["expected"] = *detail.expected;
-          if (detail.actual.has_value())
-            row["actual"] = *detail.actual;
-          if (detail.expected_text.has_value())
-            row["expected"] = *detail.expected_text;
-          if (detail.actual_text.has_value())
-            row["actual"] = *detail.actual_text;
-          if (detail.intentional.has_value())
-            row["intentional"] = *detail.intentional;
-          return row;
-        } else if constexpr (std::is_same_v<detail_t, yy_storage::episode_fsck_issue>) {
-          auto row = yy_storage::render_episode_fsck_issue(detail);
-          if (scope != storage_fsck_scope::Episode)
-            row["projection"] = "episode-manifest";
-          return row;
-        } else if constexpr (std::is_same_v<detail_t, episode_frame_verification_issue>) {
-          return render_episode_frame_verification_issue(detail);
-        } else {
-          nlohmann::json row = {{"code", issue.code}, {"projection", detail.name}, {"path", detail.path}};
-          if (issue.code == "projection_absent") {
-            row["reason"] = "projection is derived and can be rebuilt";
-          } else {
-            row["drift"] = nlohmann::json::array();
-            for (const auto &drift : detail.verification.drift) {
-              row["drift"].push_back({{"table", drift.table},
-                                      {"projection_rows", drift.projection_rows},
-                                      {"journal_distinct", drift.journal_distinct},
-                                      {"reason", drift.reason},
-                                      {"projection_digest", drift.projection_digest},
-                                      {"journal_digest", drift.journal_digest}});
-            }
-          }
-          return row;
-        }
-      },
-      issue.detail);
-  return rendered;
-}
-
-nlohmann::json render_storage_fsck_result(const storage_fsck_result &result) {
-  nlohmann::json errors = nlohmann::json::array();
-  nlohmann::json warnings = nlohmann::json::array();
-  for (const auto &issue : result.issues) {
-    auto rendered = render_storage_fsck_issue(issue, result.scope);
-    (issue.severity == "error" ? errors : warnings).push_back(std::move(rendered));
-  }
-  nlohmann::json checked = {{"sources", result.checked.sources},
-                            {"manifests", result.checked.manifests},
-                            {"accepted_ranges", result.checked.accepted_ranges},
-                            {"source_records", result.checked.source_records},
-                            {"projection_indexes", result.checked.projection_indexes},
-                            {"orphan_payloads", result.checked.orphan_payloads},
-                            {"episode_manifest_records", result.checked.episode_manifest_records},
-                            {"episodes", result.checked.episodes}};
-  if (result.scope == storage_fsck_scope::Episode) {
-    checked["payloads"] = 0;
-    checked["schemas"] = 0;
-    checked["sqlite_projection_rows"] = 0;
-  } else {
-    checked["manifest_entries"] = result.checked.manifest_entries;
-    checked["payloads"] = result.checked.payloads;
-    checked["entries_documents"] = result.checked.entries_documents;
-  }
-  if (result.frame_verification.has_value())
-    checked["episode_frames_verified"] = result.checked.episode_frames_verified;
-
-  nlohmann::json report = {{"ok", result.ok},
-                           {"status", result.status},
-                           {"scope", storage_fsck_scope_name(result.scope)},
-                           {"degraded", result.degraded},
-                           {"errors", std::move(errors)},
-                           {"warnings", std::move(warnings)},
-                           {"checked", std::move(checked)},
-                           {"episode_manifest", yy_storage::render_episode_fsck_result(result.episode_manifest)}};
-  if (result.scope == storage_fsck_scope::Episode) {
-    report["episode_id"] = result.episode_id.has_value() ? nlohmann::json(*result.episode_id) : nlohmann::json(nullptr);
-    const auto &projection = result.projections.front();
-    report["episode_projection"] = projection_verification_json(projection.verification);
-    report["projections"] = nlohmann::json::array({{{"name", "episode-manifest"},
-                                                    {"schema", yy_storage::EPISODE_MANIFEST_SCHEMA_V1},
-                                                    {"authority", "yijinjing-journal"},
-                                                    {"path", "journal/system/storage/episode-manifest/live/*.journal"},
-                                                    {"rebuildable", false}},
-                                                   {{"name", "episode-manifest-sqlite"},
-                                                    {"schema", EPISODE_MANIFEST_PROJECTION_SCHEMA_V1},
-                                                    {"authority", "yijinjing-journal"},
-                                                    {"path", "storage/projections/episode-manifest.sqlite"},
-                                                    {"rebuildable", true}}});
-    report["repair_policy"] = {{"mode", "plan-fetch-apply"},
-                               {"auto_repair", false},
-                               {"destructive", false},
-                               {"projection_rebuild",
-                                {{"authority", "yijinjing-journal"},
-                                 {"projection", "sqlite"},
-                                 {"operation", "episode_projection_rebuild"},
-                                 {"rebuildable", true}}}};
-    if (result.qualification.has_value())
-      report["qualification"] = episode_qualification_json(*result.qualification);
-  } else {
-    report["source_id"] = result.source_id.has_value() ? nlohmann::json(*result.source_id) : nlohmann::json(nullptr);
-    report["authority"] = result.authority;
-    report["projections"] = nlohmann::json::array();
-    for (const auto &projection : result.projections)
-      report["projections"].push_back(projection_status_json(projection));
-  }
-  return report;
-}
-
-nlohmann::json render_storage_repair_plan_result(const storage_repair_plan_result &result) {
-  const auto render_subject = [](nlohmann::json &candidate, const storage_repair_subject &subject) {
-    if (subject.episode_id.has_value())
-      candidate["episode_id"] = *subject.episode_id;
-    if (subject.dependency_episode_id.has_value())
-      candidate["dependency_episode_id"] = *subject.dependency_episode_id;
-    if (subject.frame_uid.has_value())
-      candidate["frame_uid"] = *subject.frame_uid;
-    if (subject.dependent_frame_uid.has_value())
-      candidate["dependent_frame_uid"] = *subject.dependent_frame_uid;
-    if (subject.ref_id.has_value())
-      candidate["ref_id"] = *subject.ref_id;
-    if (subject.ref_hash.has_value())
-      candidate["ref_hash"] = *subject.ref_hash;
-    if (subject.source_id.has_value())
-      candidate["source_id"] = *subject.source_id;
-    if (subject.subject.has_value())
-      candidate["subject"] = *subject.subject;
-    if (subject.state.has_value())
-      candidate["state"] = *subject.state;
-    if (subject.path.has_value())
-      candidate["path"] = *subject.path;
-    if (subject.payload_hash.has_value())
-      candidate["payload_hash"] = *subject.payload_hash;
-  };
-  nlohmann::json candidates = nlohmann::json::array();
-  for (const auto &item : result.candidates) {
-    nlohmann::json candidate = {{"code", item.code},
-                                {"issue_code", item.issue_code},
-                                {"kind", item.kind},
-                                {"role", item.role},
-                                {"action", item.action},
-                                {"suggested_action", item.action},
-                                {"safe_to_apply", item.safe_to_apply},
-                                {"requires", item.required_inputs},
-                                {"warning", render_storage_fsck_issue(item.issue, result.scope)}};
-    render_subject(candidate, item.subject);
-    candidates.push_back(std::move(candidate));
-  }
-  nlohmann::json unsupported = nlohmann::json::array();
-  for (const auto &issue : result.unsupported)
-    unsupported.push_back(render_storage_fsck_issue(issue, result.scope));
-  return {{"ok", result.ok},
-          {"schema", "kungfu.storage.repair-plan/v1"},
-          {"scope", storage_fsck_scope_name(result.scope)},
-          {"source_id", result.source_id.has_value() ? nlohmann::json(*result.source_id) : nlohmann::json(nullptr)},
-          {"episode_id", result.episode_id.has_value() ? nlohmann::json(*result.episode_id) : nlohmann::json(nullptr)},
-          {"dry_run", result.dry_run},
-          {"plan_only", result.plan_only},
-          {"status", result.status},
-          {"degraded", result.degraded},
-          {"candidate_count", candidates.size()},
-          {"candidates", std::move(candidates)},
-          {"unsupported", std::move(unsupported)},
-          {"fsck", render_storage_fsck_result(result.fsck)},
-          {"notes", result.notes}};
-}
-
-storage_status_request parse_storage_status_request(const storage_service_options &options) {
-  return {options.runtime_dir, options.provider, options.provider_config_source, options.source_id};
-}
-
-storage_gc_plan_request parse_storage_gc_plan_request(const storage_service_options &options) {
-  return {options.runtime_dir, options.provider, options.source_id, options.dry_run};
-}
-
-nlohmann::json render_storage_gc_plan_result(const storage_gc_plan_result &result) {
-  nlohmann::json candidates = nlohmann::json::array();
-  for (const auto &candidate : result.candidates) {
-    candidates.push_back({{"payload_hash", candidate.payload_hash},
-                          {"path", candidate.uri},
-                          {"bytes", candidate.bytes},
-                          {"safe_to_delete", candidate.safe_to_delete}});
-  }
-  return {{"ok", result.ok},
-          {"scope", result.scope},
-          {"source_id", result.source_id.has_value() ? nlohmann::json(*result.source_id) : nlohmann::json(nullptr)},
-          {"dry_run", result.dry_run},
-          {"payloads_scanned", result.payloads_scanned},
-          {"referenced_payloads", result.referenced_payloads},
-          {"candidate_count", result.candidates.size()},
-          {"candidate_bytes", result.candidate_bytes},
-          {"candidates", std::move(candidates)},
-          {"notes", result.notes}};
-}
-
-storage_rebuild_index_request parse_storage_rebuild_index_request(const storage_service_options &options) {
-  return {options.runtime_dir, options.source_id, options.dry_run};
-}
-
-nlohmann::json render_storage_rebuild_index_result(const storage_rebuild_index_result &result) {
-  nlohmann::json projections = nlohmann::json::array();
-  for (const auto &action : result.projections) {
-    auto rendered = std::visit(
-        [](const auto &detail) {
-          using detail_t = std::decay_t<decltype(detail)>;
-          nlohmann::json row = {{"ok", detail.ok},
-                                {"schema", detail.schema},
-                                {"runtime_dir", detail.runtime_dir},
-                                {"authority", detail.authority}};
-          if constexpr (std::is_same_v<detail_t, storage_projection_verify_result>) {
-            row["status"] = detail.status;
-            row["projection_present"] = detail.projection_present;
-            if (!detail.note.empty()) {
-              row["note"] = detail.note;
-            }
-            if (detail.projection_present) {
-              row["degraded"] = detail.degraded;
-              row["drift"] = nlohmann::json::array();
-              for (const auto &drift : detail.drift) {
-                row["drift"].push_back({{"table", drift.table},
-                                        {"projection_rows", drift.projection_rows},
-                                        {"journal_distinct", drift.journal_distinct},
-                                        {"reason", drift.reason},
-                                        {"projection_digest", drift.projection_digest},
-                                        {"journal_digest", drift.journal_digest}});
-              }
-              row["rows"] = nlohmann::json::object();
-              for (const auto &count : detail.rows) {
-                row["rows"][count.table] = count.count;
-              }
-              row["journal_distinct"] = nlohmann::json::object();
-              for (const auto &count : detail.journal_distinct) {
-                row["journal_distinct"][count.table] = count.count;
-              }
-            }
-          } else {
-            row["projection"] = detail.projection;
-            row["sqlite_path"] = detail.sqlite_path;
-            row["rows"] = nlohmann::json::object();
-            for (const auto &count : detail.rows) {
-              row["rows"][count.table] = count.count;
-            }
-            row["journal_records"] = nlohmann::json::object();
-            for (const auto &count : detail.journal_records) {
-              row["journal_records"][count.table] = count.count;
-            }
-          }
-          return row;
-        },
-        action.detail);
-    rendered["name"] = action.name;
-    rendered["dry_run"] = action.dry_run;
-    rendered["written"] = action.written;
-    if (action.dry_run) {
-      rendered["would_write"] = action.would_write;
-    }
-    projections.push_back(std::move(rendered));
-  }
-  nlohmann::json errors = nlohmann::json::array();
-  for (const auto &error : result.errors) {
-    nlohmann::json row = {{"code", error.code}};
-    if (error.projection.has_value()) {
-      row["projection"] = *error.projection;
-    }
-    if (error.source_id.has_value()) {
-      row["source_id"] = *error.source_id;
-    }
-    errors.push_back(std::move(row));
-  }
-  return {{"ok", result.ok},
-          {"scope", result.scope},
-          {"source_id", result.source_id.has_value() ? nlohmann::json(*result.source_id) : nlohmann::json(nullptr)},
-          {"authority", result.authority},
-          {"rebuilt_from", result.rebuilt_from},
-          {"projections", std::move(projections)},
-          {"dry_run", result.dry_run},
-          {"would_write", result.would_write},
-          {"written", result.written},
-          {"sources_rebuilt", result.sources_rebuilt},
-          {"errors", std::move(errors)}};
-}
-
-storage_compact_plan_request parse_storage_compact_plan_request(const storage_service_options &options) {
-  return {options.runtime_dir, options.provider, options.source_id, options.dry_run};
-}
-
-nlohmann::json render_storage_compact_plan_result(const storage_compact_plan_result &result) {
-  nlohmann::json retained_manifests = nlohmann::json::array();
-  for (const auto &manifest : result.retained_manifests) {
-    retained_manifests.push_back({{"source_id", manifest.source_id},
-                                  {"manifest_id", manifest.manifest_id},
-                                  {"entries", manifest.entries},
-                                  {"sync_root",
-                                   {{"schema", yy_storage::SYNC_ROOT_SCHEMA_V1},
-                                    {"scope", yy_storage::SYNC_ROOT_SCOPE_SOURCE_IMPORT_MANIFEST},
-                                    {"proof", yy_storage::SYNC_ROOT_PROOF_LINEAR_CHAIN_V1},
-                                    {"algorithm", manifest.sync_root.algorithm},
-                                    {"value", manifest.sync_root.value},
-                                    {"entry_count", manifest.entries},
-                                    {"initial", yy_storage::SYNC_ROOT_INITIAL_SHA256},
-                                    {"ordering",
-                                     {{"policy", yy_storage::SYNC_ROOT_ORDERING_POLICY_MANIFEST_ENTRY_SORT_V1},
-                                      {"fields", nlohmann::json::array({"kind", "source_id", "source_path"})}}}}}});
-  }
-  nlohmann::json unsupported = nlohmann::json::array();
-  for (const auto &action : result.unsupported) {
-    unsupported.push_back({{"name", action.name}, {"reason", action.reason}});
-  }
-  return {
-      {"ok", result.ok},
-      {"scope", result.scope},
-      {"source_id", result.source_id.has_value() ? nlohmann::json(*result.source_id) : nlohmann::json(nullptr)},
-      {"dry_run", result.dry_run},
-      {"retained_manifests", std::move(retained_manifests)},
-      {"rebuild_index", render_storage_rebuild_index_result(result.rebuild_index)},
-      {"gc", render_storage_gc_plan_result(result.gc)},
-      {"projection_compact",
-       {{"name", result.projection_compact.name},
-        {"path", result.projection_compact.path},
-        {"action", result.projection_compact.action},
-        {"dry_run", result.projection_compact.dry_run},
-        {"rebuildable", result.projection_compact.rebuildable}}},
-      {"unsupported", std::move(unsupported)},
-      {"notes", result.notes},
-  };
-}
-
-nlohmann::json render_storage_status_result(const storage_status_result &result) {
-  const auto range_json = [](const storage_time_range &range) {
-    nlohmann::json rendered = nlohmann::json::object();
-    if (!range.since.empty()) {
-      rendered["since"] = range.since;
-    }
-    if (!range.until.empty()) {
-      rendered["until"] = range.until;
-    }
-    return rendered;
-  };
-  const auto sync_root_json = [](const storage_sync_root_view &root) {
-    return nlohmann::json{{"algorithm", root.algorithm}, {"value", root.value}};
-  };
-  const auto proof_root_json = [&sync_root_json](const storage_sync_root_view &root, uint64_t entry_count) {
-    auto rendered = sync_root_json(root);
-    rendered["schema"] = yy_storage::SYNC_ROOT_SCHEMA_V1;
-    rendered["scope"] = yy_storage::SYNC_ROOT_SCOPE_SOURCE_IMPORT_MANIFEST;
-    rendered["proof"] = yy_storage::SYNC_ROOT_PROOF_LINEAR_CHAIN_V1;
-    rendered["entry_count"] = entry_count;
-    rendered["initial"] = yy_storage::SYNC_ROOT_INITIAL_SHA256;
-    rendered["ordering"] = {{"policy", yy_storage::SYNC_ROOT_ORDERING_POLICY_MANIFEST_ENTRY_SORT_V1},
-                            {"fields", nlohmann::json::array({"kind", "source_id", "source_path"})}};
-    return rendered;
-  };
-  const auto accepted_range_json = [&range_json, &proof_root_json](const storage_accepted_range_view &range) {
-    return nlohmann::json{{"schema", yy_storage::STORAGE_ACCEPTED_RANGE_SCHEMA_V1},
-                          {"source_id", range.source_id},
-                          {"manifest_id", range.manifest_id},
-                          {"range", range_json(range.range)},
-                          {"source_head", range.source_head},
-                          {"sync_root", proof_root_json(range.sync_root, range.entry_count)},
-                          {"entry_count", range.entry_count},
-                          {"status", range.status}};
-  };
-  const auto source_json = [](const storage_source_registry_view &source) {
-    nlohmann::json rendered = {{"schema", yy_storage::SOURCE_REGISTRY_SCHEMA_V1},
-                               {"source_uid", source.source_uid},
-                               {"source_id", source.source_id},
-                               {"registered", source.registered},
-                               {"record_count", source.record_count},
-                               {"accepted_range_count", source.accepted_range_count}};
-    if (source.registered) {
-      rendered["kind"] = *source.kind;
-      rendered["coordinate"] = *source.coordinate;
-      rendered["head"] = *source.head;
-      rendered["location_uid"] = *source.location_uid;
-      rendered["register_time"] = *source.register_time;
-    }
-    if (source.current_range.has_value()) {
-      const auto &range = *source.current_range;
-      rendered["current_range"] = {{"first_frame_uid", range.first_frame_uid},
-                                   {"last_frame_uid", range.last_frame_uid},
-                                   {"since", range.since},
-                                   {"until", range.until}};
-      rendered["inventory_hash"] = {{"algorithm", source.inventory_hash->algorithm},
-                                    {"value", source.inventory_hash->value}};
-      rendered["update_time"] = *source.update_time;
-    }
-    return rendered;
-  };
-
-  nlohmann::json sources = nlohmann::json::array();
-  for (const auto &source : result.sources) {
-    sources.push_back(source_json(source));
-  }
-  nlohmann::json projections = nlohmann::json::array();
-  for (const auto &projection : result.projections) {
-    projections.push_back(projection_status_json(projection));
-  }
-  nlohmann::json source_status = nlohmann::json::array();
-  for (const auto &status : result.source_status) {
-    nlohmann::json row = {{"source_id", status.source_id}, {"ok", status.ok}, {"authority", result.authority}};
-    if (!status.ok) {
-      row["reason"] = *status.reason;
-      row["source"] = source_json(status.source);
-      source_status.push_back(std::move(row));
-      continue;
-    }
-    row["manifest_id"] = *status.manifest_id;
-    row["source_type"] = *status.source_type;
-    row["source_head"] = *status.source_head;
-    row["accepted_ranges"] = nlohmann::json::array({accepted_range_json(*status.accepted_range)});
-    if (status.accepted_cursor.has_value()) {
-      const auto &cursor = *status.accepted_cursor;
-      row["accepted_cursor"] = {{"schema", yy_storage::STORAGE_CHANNEL_CURSOR_SCHEMA_V1},
-                                {"source_id", cursor.source_id},
-                                {"manifest_id", cursor.manifest_id},
-                                {"source_head", cursor.source_head},
-                                {"range", range_json(cursor.range)},
-                                {"sync_root", sync_root_json(cursor.sync_root)},
-                                {"entry_count", cursor.entry_count}};
-    } else {
-      row["accepted_cursor"] = nullptr;
-    }
-    row["sync_root"] = proof_root_json(*status.sync_root, status.entries);
-    row["entries"] = status.entries;
-    row["payload_inventory"] = status.payload_inventory;
-    row["schema_inventory"] = status.schema_inventory;
-    const auto &record = *status.source_record;
-    row["source_record"] = {{"schema", yy_storage::STORAGE_SOURCE_RECORD_SCHEMA_V1},
-                            {"source_id", record.source_id},
-                            {"type", record.source_type},
-                            {"kind", record.kind},
-                            {"coordinate", record.coordinate},
-                            {"current_head",
-                             {{"head", record.source_head},
-                              {"range", range_json(record.range)},
-                              {"inventory_hash", record.inventory_hash}}},
-                            {"accepted_ranges", nlohmann::json::array({accepted_range_json(record.accepted_range)})},
-                            {"last_manifest_id", record.manifest_id},
-                            {"updated_at", ""}};
-    source_status.push_back(std::move(row));
-  }
-
-  return {{"ok", result.ok},
-          {"backend", result.backend},
-          {"provider", result.provider},
-          {"provider_config_source", result.provider_config_source},
-          {"provider_runtime", provider_runtime_json(result.provider_runtime)},
-          {"provider_cache", provider_cache_json(result.provider_cache)},
-          {"scope", result.scope},
-          {"source_id", result.source_id.has_value() ? nlohmann::json(*result.source_id) : nlohmann::json(nullptr)},
-          {"authority", result.authority},
-          {"sources", std::move(sources)},
-          {"source_count", result.sources.size()},
-          {"projections", std::move(projections)},
-          {"source_status", std::move(source_status)}};
-}
-
-storage_service_options parse_storage_service_options(const std::string &runtime_dir, const nlohmann::json &options) {
-  storage_service_options parsed;
-  parsed.runtime_dir = runtime_dir;
-  const auto selected_provider = select_provider_for_runtime(runtime_dir, text_or(options, "provider"));
-  parsed.provider = selected_provider.name;
-  parsed.provider_config_source = selected_provider.source;
-  parsed.scope = text_or(options, "scope", "all");
-  parsed.source_id = text_or(options, "source_id");
-  parsed.dry_run = bool_or(options, "dry_run", true);
-  parsed.verify = bool_or(options, "verify", true);
-  parsed.range = object_or_empty(options, "range");
-  parsed.artifact_uri = text_or(options, "artifact_uri");
-  parsed.bundle = object_or_empty(options, "bundle");
-  parsed.manifest = object_or_empty(options, "manifest");
-  parsed.operation_options = options;
-  parsed.query_definition = object_or_empty(options, "definition");
-  parsed.query = text_or(options, "query", "entries");
-  parsed.kind = text_or(options, "kind");
-  parsed.episode_id = uint64_or(options, "episode_id");
-  parsed.limit = uint64_or(options, "limit", 100);
-  return parsed;
-}
-
-nlohmann::json make_storage_service_request(const std::string &operation, const std::string &runtime_dir,
-                                            const nlohmann::json &options) {
-  const auto parsed_operation = parse_storage_operation(operation);
-  const auto parsed_options = parse_storage_service_options(runtime_dir, options);
-  auto request = make_request(parsed_operation, parsed_options);
-  if (parsed_operation == storage_operation::Query) {
-    request["query"] = parsed_options.query;
-    request["kind"] = parsed_options.kind.empty() ? nlohmann::json(nullptr) : nlohmann::json(parsed_options.kind);
-    request["limit"] = parsed_options.limit;
-  } else if (parsed_operation == storage_operation::QueryPlan || parsed_operation == storage_operation::FactQuery ||
-             parsed_operation == storage_operation::FactChangelog ||
-             parsed_operation == storage_operation::SavedQueryCatalog) {
-    request["definition"] = parsed_options.query_definition;
-    request["action"] = text_or(parsed_options.operation_options, "action");
-  } else if (parsed_operation == storage_operation::Layout) {
-    request["runtime_home"] = runtime_home_path(parsed_options).string();
-    request["runtime_home_source"] = runtime_home_source(parsed_options);
-    request["config_home"] = optional_absolute_path(parsed_options.operation_options, "config_home");
-  }
-  if (parsed_options.episode_id != 0) {
-    request["episode_id"] = parsed_options.episode_id;
-  }
-  if (parsed_operation == storage_operation::EpisodeList) {
-    request["location_uid"] = uint64_or(options, "location_uid");
-    request["limit"] = parsed_options.limit;
-  } else if (parsed_operation == storage_operation::EpisodeInspect) {
-    request["episode_id"] = uint64_or(options, "episode_id");
-  } else if (storage_operation_name(parsed_operation).rfind("episode_", 0) == 0) {
-    request["episode"] = parsed_options.operation_options;
-  }
-  return request;
-}
-
-nlohmann::json run_storage_service_operation(const std::string &operation, const std::string &runtime_dir,
-                                             const nlohmann::json &options) {
-  const auto parsed_operation = parse_storage_operation(operation);
-  if (parsed_operation == storage_operation::BackendStatus || parsed_operation == storage_operation::BackendSwitch ||
-      parsed_operation == storage_operation::BackendRollback) {
-    storage_service_options backend_options{};
-    backend_options.runtime_dir = runtime_dir;
-    backend_options.operation_options = options;
-    return dispatch_backend_operation(parsed_operation, backend_options);
-  }
-  return detail::dispatch_json_edge_operation(parsed_operation, parse_storage_service_options(runtime_dir, options));
-}
-
-nlohmann::json accept_storage_manifest(const std::string &runtime_dir, const nlohmann::json &manifest) {
-  return accept_storage_manifest_impl(runtime_dir, manifest);
-}
-
-nlohmann::json load_storage_latest_manifest(const std::string &runtime_dir, const std::string &source_id) {
-  const auto provider = shared_provider(runtime_dir);
-  return load_latest_manifest_impl(runtime_dir, *provider, source_id);
-}
-
-nlohmann::json export_storage_records(const std::string &runtime_dir, const std::string &source_id,
-                                      const nlohmann::json &range) {
-  const auto provider = shared_provider(runtime_dir);
-  const auto manifest = load_latest_manifest_impl(runtime_dir, *provider, source_id);
-  if (manifest.is_null()) {
-    throw std::runtime_error("manifest not found: " + source_id);
-  }
-  nlohmann::json records = nlohmann::json::array();
-  for (const auto &entry : entries_for_manifest(manifest, range)) {
-    auto row = entry;
-    row["scope"] = text_or(manifest, "scope");
-    row["manifest_id"] = text_or(manifest, "manifest_id");
-    row["storage_source_id"] = source_id;
-    row["source_type"] = text_or(manifest, "source_type");
-    row["source_head"] = text_or(manifest, "source_head");
-    const auto state = text_or(entry, "payload_state", PAYLOAD_STATE_PRESENT);
-    if (state == PAYLOAD_STATE_REDACTED || state == PAYLOAD_STATE_ABSENT) {
-      row["payload"] = nullptr;
-    } else if (state != PAYLOAD_STATE_PRESENT) {
-      auto [payload, error] = load_payload_impl(*provider, entry);
-      row["payload"] = error.empty() ? payload : nlohmann::json(nullptr);
-    } else {
-      auto [payload, error] = load_payload_impl(*provider, entry);
-      if (!error.empty()) {
-        throw std::runtime_error(error + ": " + text_or(entry, "kind") + ":" + text_or(entry, "source_id"));
-      }
-      row["payload"] = payload;
-    }
-    records.push_back(row);
-  }
-  std::sort(records.begin(), records.end(), [](const nlohmann::json &lhs, const nlohmann::json &rhs) {
-    return std::make_tuple(text_or(lhs, "kind"), text_or(lhs, "source_id"), text_or(lhs, "source_path")) <
-           std::make_tuple(text_or(rhs, "kind"), text_or(rhs, "source_id"), text_or(rhs, "source_path"));
-  });
-  return records;
-}
-
-std::string write_storage_payload_bytes(const std::string &runtime_dir, const std::string &digest,
-                                        const std::string &raw) {
-  const auto error = yy_storage::verify_payload_ref(raw, digest, raw.size(), yy_storage::CONTENT_HASH_ALGORITHM_SHA256);
-  if (!error.empty()) {
-    throw std::invalid_argument("storage_payload_invalid: " + error);
-  }
-  const auto provider = shared_provider(runtime_dir);
-  provider->write_payload(digest, raw);
-  return payload_uri_for(provider->name(), runtime_dir, digest);
-}
-
-namespace detail {
-
-nlohmann::json content_result_json(const yy_storage::content_store_result &result) {
-  return {{"ok", result.ok()},
-          {"error", yy_storage::content_store_error_name(result.error)},
-          {"hash", {{"algorithm", result.hash.algorithm}, {"value", result.hash.value}}},
-          {"byte_length", result.byte_length},
-          {"existed", result.existed},
-          {"message", result.message}};
-}
-
-} // namespace detail
-
-namespace {
-
-bool parse_content_hash_text(const std::string &text, yy_storage::content_hash &hash, std::string &message) {
-  try {
-    hash = text.find(':') != std::string::npos ? yy_storage::parse_content_hash(text)
-                                               : yy_storage::make_content_hash(text);
-    return true;
-  } catch (const std::exception &e) {
-    message = e.what();
-    return false;
-  }
-}
-
-nlohmann::json invalid_content_hash_json(const std::string &message) {
-  yy_storage::content_store_result result{};
-  result.error = yy_storage::content_store_error::InvalidArgument;
-  result.message = message;
-  return content_result_json(result);
-}
-
-} // namespace
-
-// KF-ADR-019f86da-4f90-738c-b372-e509976f69ff routes one immutable content-store contract through the runtime-selected
-// provider.
-nlohmann::json content_store_put_if_absent(const std::string &runtime_dir, const std::string &content_namespace,
-                                           const std::string &raw, const std::string &expected_hash) {
-  yy_storage::content_hash expected{};
-  if (!expected_hash.empty()) {
-    std::string message;
-    if (!parse_content_hash_text(expected_hash, expected, message)) {
-      return invalid_content_hash_json(message);
-    }
-  }
-  const auto provider = shared_provider(runtime_dir);
-  return content_result_json(provider->content_store().put_if_absent(content_namespace, raw, expected));
-}
-
-bool content_store_has(const std::string &runtime_dir, const std::string &content_namespace,
-                       const std::string &content_hash_text) {
-  yy_storage::content_hash hash{};
-  std::string message;
-  if (!parse_content_hash_text(content_hash_text, hash, message)) {
-    return false;
-  }
-  const auto provider = shared_provider(runtime_dir);
-  return provider->content_store().has(content_namespace, hash);
-}
-
-nlohmann::json content_store_verify(const std::string &runtime_dir, const std::string &content_namespace,
-                                    const std::string &content_hash_text) {
-  yy_storage::content_hash hash{};
-  std::string message;
-  if (!parse_content_hash_text(content_hash_text, hash, message)) {
-    return invalid_content_hash_json(message);
-  }
-  const auto provider = shared_provider(runtime_dir);
-  return content_result_json(provider->content_store().verify(content_namespace, hash));
-}
-
-std::string content_store_get(const std::string &runtime_dir, const std::string &content_namespace,
-                              const std::string &content_hash_text) {
-  yy_storage::content_hash hash{};
-  std::string message;
-  if (!parse_content_hash_text(content_hash_text, hash, message)) {
-    throw std::invalid_argument("content_store_get: " + message);
-  }
-  const auto provider = shared_provider(runtime_dir);
-  auto result = provider->content_store().get(content_namespace, hash);
-  if (!result.ok()) {
-    throw std::runtime_error(std::string("content_store_get_failed: ") +
-                             yy_storage::content_store_error_name(result.error) +
-                             (result.message.empty() ? "" : ": " + result.message));
-  }
-  return std::move(result.bytes);
-}
-
-nlohmann::json content_store_capabilities(const std::string &runtime_dir) {
-  const auto provider = shared_provider(runtime_dir);
-  const auto caps = provider->content_store().capabilities();
-  return {{"provider", provider->name()},
-          {"profile", caps.profile},
-          {"hash_algorithm", caps.hash_algorithm},
-          {"max_object_size", caps.max_object_size},
-          {"atomic_put_if_absent", caps.atomic_put_if_absent},
-          {"verified_reads", caps.verified_reads},
-          {"durability", caps.durability},
-          {"visibility", caps.visibility},
-          {"concurrency", caps.concurrency}};
-}
-
-nlohmann::json storage_service_capabilities() {
-  const auto provider = select_provider({});
-  return {
-      {"schema", RUNTIME_STORAGE_SERVICE_SCHEMA_V1},
-      {"owner", RUNTIME_STORAGE_SERVICE_OWNER},
-      {"operations", storage_operation_names()},
-      {"backend", provider.name},
-      {"provider", provider.name},
-      {"provider_config_source", provider.source},
-      {"providers", nlohmann::json::array({
-                        {{"name", PROVIDER_FILE},
-                         {"available", provider_available(PROVIDER_FILE)},
-                         {"default", provider.name == PROVIDER_FILE},
-                         {"selected", provider.name == PROVIDER_FILE},
-                         {"layout", provider_layout_json(provider_layout_for(PROVIDER_FILE))},
-                         {"runtime", provider_runtime_json(provider_runtime_for(PROVIDER_FILE))}},
-                        {{"name", PROVIDER_ROCKSDB},
-                         {"available", provider_available(PROVIDER_ROCKSDB)},
-                         {"default", provider.name == PROVIDER_ROCKSDB},
-                         {"selected", provider.name == PROVIDER_ROCKSDB},
-                         {"layout", provider_layout_json(provider_layout_for(PROVIDER_ROCKSDB))},
-                         {"runtime", provider_runtime_json(provider_runtime_for(PROVIDER_ROCKSDB))}},
-                    })},
-      {"backend_authority", backend_authority_capability_json()},
-      {"projections",
-       nlohmann::json::array(
-           {{{"name", "source-registry"},
-             {"schema", yy_storage::SOURCE_REGISTRY_SCHEMA_V1},
-             {"authority", "yijinjing-journal"},
-             {"path", "journal/system/storage/source-registry/live/*.journal"},
-             {"rebuildable", false}},
-            {{"name", PROJECTION_SOURCE_REGISTRY},
-             {"schema", SOURCE_REGISTRY_PROJECTION_SCHEMA_V1},
-             {"authority", "derived"},
-             {"path", "storage/projections/source-registry.sqlite"},
-             {"rebuildable", true}},
-            {{"name", "manifest-catalog"},
-             {"schema", yy_storage::MANIFEST_CATALOG_SCHEMA_V1},
-             {"authority", "yijinjing-journal"},
-             {"path", "journal/system/storage/manifest-catalog/live/*.journal"},
-             {"export_schema", yy_storage::STORAGE_EXPORT_BUNDLE_SCHEMA_V1},
-             {"rebuildable", false}},
-            {{"name", PROJECTION_MANIFEST_CATALOG},
-             {"schema", MANIFEST_CATALOG_PROJECTION_SCHEMA_V1},
-             {"authority", "derived"},
-             {"path", "storage/projections/manifest-catalog.sqlite"},
-             {"rebuildable", true}},
-            {{"name", "episode-manifest"},
-             {"schema", yy_storage::EPISODE_MANIFEST_SCHEMA_V1},
-             {"authority", "yijinjing-journal"},
-             {"path", "journal/system/storage/episode-manifest/live/*.journal"},
-             {"query_tables", nlohmann::json::array({"episodes", "episode_records", "episode_frames", "episode_refs"})},
-             {"export_schema", "kungfu.storage.episode-bundle/v1"},
-             {"rebuildable", false}},
-            {{"name", "domain-fact-admission"},
-             {"schema", facts::DOMAIN_FACT_EVENT_SCHEMA_V1},
-             {"authority", "yijinjing-journal"},
-             {"path", "journal/system/facts/admission/live/*.journal"},
-             {"contract", facts::fact_contract_json()},
-             {"rebuildable", false}}})},
-      {"notes",
-       nlohmann::json::array({
-           "The runtime storage service surface and providers are owned by libkungfu.",
-           "Provider selection is an implementation option; product semantics remain storage-service operations.",
-           "Python and Node should remain binding, CLI, or UI layers over this service.",
-       })},
-  };
-}
 
 } // namespace kungfu::runtime::storage_service_api
